// src/api/generated/model/mgmtTarget.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */
import type { MgmtPollStatus } from './mgmtPollStatus';
import type { Links } from './links';

/**
 * **_links**:
* **assignedDS** - Links to assigned distribution sets
* **installedDS** - Links to installed distribution sets
* **attributes** - Links to attributes of the target
* **actions** - Links to actions of the target
* **metadata** - List of metadata
* **targetType** - The link to the target type
* **autoConfirm** - The link to the detailed auto confirm state

 */
export interface MgmtTarget {
  /** Entity was originally created by (User, AMQP-Controller, anonymous etc.) */
  readonly createdBy?: string;
  /** Entity was originally created at (timestamp UTC in milliseconds) */
  readonly createdAt?: number;
  /** Entity was last modified by (User, AMQP-Controller, anonymous etc.) */
  readonly lastModifiedBy?: string;
  /** Entity was last modified at (timestamp UTC in milliseconds) */
  readonly lastModifiedAt?: number;
  name: string;
  description?: string;
  /** Controller ID */
  controllerId: string;
  /** If the target is in sync */
  updateStatus?: string;
  /** Timestamp of the last controller request */
  lastControllerRequestAt?: number;
  /** Install timestamp */
  installedAt?: number;
  /** Last known IP address of the target. Only presented if IP address of the target itself is known (connected directly through DDI API) */
  ipAddress?: string;
  /** The last known address URI of the target. Includes information of the target is connected either directly (DDI) through HTTP or indirectly (DMF) through amqp. */
  address?: string;
  pollStatus?: MgmtPollStatus;
  /** Pre-Shared key that allows targets to authenticate at Direct Device Integration API if enabled in the tenant settings */
  securityToken?: string;
  /** Request re-transmission of target attributes */
  requestAttributes?: boolean;
  /** ID of the target type */
  targetType?: number;
  /** Name of the target type */
  targetTypeName?: string;
  /** Present if user consent flow active. Indicates if auto-confirm is active */
  autoConfirmActive?: boolean;
  _links?: Links;
}