// src/api/generated/model/mgmtTag.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */
import type { Links } from './links';

/**
 * **_links**:
* **assignedDistributionSets** - Links to assigned distribution sets

 */
export interface MgmtTag {
  /** Entity was originally created by (User, AMQP-Controller, anonymous etc.) */
  readonly createdBy?: string;
  /** Entity was originally created at (timestamp UTC in milliseconds) */
  readonly createdAt?: number;
  /** Entity was last modified by (User, AMQP-Controller, anonymous etc.) */
  readonly lastModifiedBy?: string;
  /** Entity was last modified at (timestamp UTC in milliseconds) */
  readonly lastModifiedAt?: number;
  name: string;
  description?: string;
  /** The technical identifier of the entity */
  id: number;
  /** The colour of the entity */
  colour?: string;
  _links?: Links;
}