// src/api/generated/model/mgmtDistributionSetAssignment.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */
import type { MgmtDistributionSetAssignmentType } from './mgmtDistributionSetAssignmentType';
import type { MgmtMaintenanceWindowRequestBody } from './mgmtMaintenanceWindowRequestBody';

export interface MgmtDistributionSetAssignment {
  id: number;
  /** Forcetime in milliseconds */
  forcetime?: number;
  /** Importance of the assignment */
  weight?: number;
  /** (Available with user consent flow active) Specifies if the confirmation by the device
is required for this action */
  confirmationRequired?: boolean;
  /** The type of the assignment */
  type?: MgmtDistributionSetAssignmentType;
  maintenanceWindow?: MgmtMaintenanceWindowRequestBody;
}